[ref]ディレクトリ構成の調整とmainの分割
