+++ model/dbope.go
@@ -1,0 +1,3 @@
+package model
+
+// 複数テーブルにわたる処理やテーブル自体への処理
