+++ view/views/scripts/common.js
